extract shared strings into standalone component, WIP
